requirements.txt
pandas
numpy
scikit-learn
pyyaml
joblib
pytest
